requirements.txt
langchain
chainlit
langchain_openai
openai
tiktoken
pymupdf
pinecone-client
langchain_pinecone
azure-search-documents
azure-identity